// backend/src/enums/errorMessages.enum.ts
export enum ErrorMessages {
  UNAUTHORIZED = 'Unauthorized',
  UNAUTHORIZED_ACCESS = 'Unauthorized access.',
  AUTH_REQUIRED = 'Authentication required. No token provided.',
  INVALID_TOKEN = 'Invalid or expired token.',
  NO_IMAGE_PROVIDED = 'No image provided.',
  AT_LEAST_ONE_PHOTO = 'At least one photo is required.',
  DUPLICATE_TITLE = 'DUPLICATE_TITLE',
  DUPLICATE_TITLE_CREATE = 'A memory with this title already exists in your vault. Consider adding to the same bulk or using a different title.',
  DUPLICATE_TITLE_UPDATE = 'Another memory with this title already exists in your vault. Please use a unique title.',
  UPLOAD_NOT_FOUND = 'Upload not found.',
  MEMORY_NOT_FOUND = 'Memory not found',
  MEMORY_NOT_FOUND_PERIOD = 'Memory not found.',
  MEMORY_NOT_FOUND_UNAUTHORIZED = 'Memory not found or unauthorized',
  MEMORY_PRIVATE = 'This memory is private.',
  ACCESS_DENIED_UNLISTED = 'Access denied. A sharing token is required for this unlisted memory.',
  FAILED_TO_REMOVE_MEMORY = 'Failed to remove memory.',
  USER_NOT_FOUND = 'User not found.',
  INVALID_CREDENTIALS = 'Invalid credentials.',
  EMAIL_IN_USE = 'Email already in use.',
  USERNAME_TAKEN = 'Username already taken.',
  INVALID_OTP = 'Invalid or expired OTP.',
  FAILED_VERIFICATION = 'Failed to update user verification status.',
  INVALID_EMAIL_PASSWORD = 'Invalid email or password.',
  ACCOUNT_NOT_VERIFIED = 'Account not verified. Please check your email.',
  USER_EMAIL_NOT_EXIST = 'User with this email does not exist.',
  INVALID_OLD_PASSWORD = 'Invalid old password.',
  FAILED_TO_UPDATE_PROFILE = 'Failed to update profile.',
  FAILED_TO_UPDATE_AVATAR = 'Failed to update avatar.',
  FORBIDDEN_OR_NOT_FOUND = 'Forbidden or not found',
  USERNAME_MIN_LENGTH = 'Username must be at least 3 characters.',
  USERNAME_MAX_LENGTH = 'Username too long.',
  INVALID_EMAIL = 'Invalid email address.',
  PASSWORD_MIN_LENGTH = 'Password must be at least 8 characters.',
  PASSWORD_UPPERCASE = 'Password must contain at least one uppercase letter.',
  PASSWORD_REQUIRED = 'Password is required.',
  OTP_LENGTH = 'OTP must be 6 digits.',
  DB_URL_MISSING = 'No DATABASE_URL set in environment variables.',
  DB_CONNECTION_ERROR = '❌ Database connection error:',
}